Update README.md
Live Streaming during conducting exam works on a handshake communication in a client-server architecture. Multiple clients send their webcam, desktop snapshots over the network to the proxized server which then stores some of them in its machine and also makes corresponding entries in mysql database and sends all image frames to the server webpage waiting for it. The server webpage then groups the data based on usernames and stream it onto the webpage.